# app/docs/openapi.json
{
  "openapi" : "3.1.0",
  "paths" : {
...
    "/preview/document/{nodeId}" : {
      "get" : {
        "parameters" : [ {
          "name" : "nodeId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "crop",
          "in" : "query",
          "schema" : {
            "type" : "boolean"
          }
        }, {
          "name" : "first_page",
          "in" : "query",
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        }, {
          "name" : "last_page",
          "in" : "query",
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        }, {
          "name" : "output_format",
          "in" : "query",
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "quality",
          "in" : "query",
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "shape",
          "in" : "query",
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "version",
          "in" : "query",
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "Cookie",
          "in" : "header",
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "ZM_AUTH_TOKEN",
          "in" : "cookie",
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : { }
              }
            }
          }
        },
        "summary" : "Preview Document",
        "tags" : [ "Preview Resource" ]
      }
    },
...
    "/preview/pdf/{nodeId}" : {
      "get" : {
        "parameters" : [ {
          "name" : "nodeId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "crop",
          "in" : "query",
          "schema" : {
            "type" : "boolean"
          }
        }, {
          "name" : "first_page",
          "in" : "query",
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        }, {
          "name" : "last_page",
          "in" : "query",
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        }, {
          "name" : "output_format",
          "in" : "query",
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "quality",
          "in" : "query",
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "shape",
          "in" : "query",
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "version",
          "in" : "query",
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "Cookie",
          "in" : "header",
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "ZM_AUTH_TOKEN",
          "in" : "cookie",
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : { }
              }
            }
          }
        },
        "summary" : "Preview Pdf",
        "tags" : [ "Preview Resource" ]
      }
    },
...
  },
...
  "info" : {
    "title" : "carbonio-files-ce-app API",
    "version" : "unversioned"
  },
  "servers" : [ {
    "url" : "http://localhost:8080",
    "description" : "Auto generated value"
  }, {
    "url" : "http://0.0.0.0:8080",
    "description" : "Auto generated value"
  } ]
}